added export message, added docs
