+++ src/docs/arc42/config.adoc
@@ -7,3 +7,8 @@
 
 // where are images located?
 :imagesdir: ./images
+
+:xrefstyle: short
+
+:uri-repo-main: https://github.com/docToolchain/drawio-intellij-plugin/blob/main/src/main
+:uri-repo-kotlin: {uri-repo-main}/kotlin/de/docs_as_co/intellij/plugin/drawio
