+++ src/main/kotlin/de/docs_as_co/intellij/plugin/drawio/editor/DrawioWebMessages.kt
@@ -21,7 +21,7 @@ sealed class OutgoingMessage {
                 val XMLSVG = "xmlsvg"
             }
 
-            val action = "setTheme"
+            val action = "export"
         }
     }
